ModalContainer の管理自体は ModalManager 側に戻す
