+++ src/lib/ModalManager.svelte
@@ -15,7 +15,6 @@
 
     _modal.classList.remove('hide');
     let $elm = document.createElement('div');
-    
 
     var instance = new ModalContainer({
       target: $elm,
@@ -74,12 +73,62 @@
 </script>
 
 <script>
-  import {onMount, onDestroy, afterUpdate} from 'svelte';
+  import {onMount, onDestroy} from 'svelte';
   let root;
 
-  afterUpdate(() => {
-    _modal = root;
-  });
+  // afterUpdate(() => {
+  //   _modal = root;
+  // });
+  let _instances = [];
+
+  // svelte-ignore unused-export-let
+  export let open = (component, props = {}) => {
+    root.classList.remove('hide');
+    let $elm = document.createElement('div');
+
+    var instance = new ModalContainer({
+      target: $elm,
+      props: {
+        component: component.default,
+        position: component.position,
+        transition: component.transition,
+        overlay: component.overlay,
+        props: {
+          ...component.defaultProps,
+          ...props,
+        },
+        destory: () => {
+          // インタンスを削除
+          instance.$destroy();
+          // リストから削除
+          let index = _instances.findIndex(item => item === instance);
+          _instances.splice(index, 1);
+          // DOM を削除
+          $elm.parentNode.removeChild($elm);
+
+          // すべてのモーダルがなくなったらモーダル自体を非表示に
+          if (root.children.length <= 0) {
+            root.classList.add('hide');
+          }
+        },
+      }
+    });
+
+    root.appendChild($elm);
+    
+    // デフォルトで modal の枠に focus しておく (ボタン連打等の対策)
+    $elm.tabIndex = '-1';
+    $elm.focus();
+
+    // リストに追加
+    _instances.push(instance);
+
+    // modal を実際に表示
+    instance.visible = true;
+
+    // modal instance を返す
+    return instance.modal;
+  };
 
   let onKeydown = (e) => {
     let current_instance = _instances[_instances.length-1];
